pyproject.toml
[tool.poetry]
name = "twitch-chatters-monitor"
version = "0.1.0"
description = ""
authors = ["ProMaker <rocker896@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
